Fix clientOnly parsing undefined and fix modals
After getting the initialCredentials, the lib was trying to persist data
using undefined variables and returning a undefined user.

Also, the variables `firstTimeLogin` and `mustResetPassword`
were never set in the clientOnly mode, so the actionsCreator related to
them was never being fired due the conditional statement requiring these
variables to be set.

The parseResponse method was refactored to allow a optional callback in
order to be reused in the applyConfig method as well.